+++ src/com/pvv/pulbet/service/EventoService.java
@@ -10,7 +10,7 @@ public interface EventoService {
 	
 	public Results<Evento> findByCriteria(EventoCriteria evento, int startIndex, int count, String idioma) throws DataException;
 	
-	public Evento findById(Integer id, String idioma) throws InstanceNotFoundException, DataException;
+	public Evento findById(Long id, String idioma) throws InstanceNotFoundException, DataException;
 	
 	public List<Long> findResultadoFinal(Long id) throws DataException;
