+++ src/com/pvv/pulbet/service/ResultadoService.java
@@ -3,7 +3,6 @@ package com.pvv.pulbet.service;
 import java.util.List;
 
 import com.pvv.pulbet.exceptions.DataException;
-import com.pvv.pulbet.exceptions.DuplicateInstanceException;
 import com.pvv.pulbet.exceptions.InstanceNotFoundException;
 import com.pvv.pulbet.model.Resultado;
 
@@ -13,7 +12,7 @@ public interface ResultadoService {
 //	
 //	public Long delete(Long id) throws InstanceNotFoundException, DataException;
 	
-	public Resultado findById(Integer id, String idioma) throws InstanceNotFoundException, DataException;
+	public Resultado findById(Long id, String idioma) throws InstanceNotFoundException, DataException;
 	
 	public List<Resultado> findByTipoResultado(Integer id, String idioma) throws DataException;
